Добавил страницу для вывода ошибок, отловил часть ошибок с авторизацей, удалением уже сущ. категорий, размеров,материалов, покрытий

+++ wood/views.py
@@ -1,10 +1,12 @@
-from django.contrib.auth.decorators import login_required
-from django.http import HttpResponse
-from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.db import IntegrityError
+from django.db import transaction
+from django.db.models import ProtectedError
+from django.http import HttpResponse
 from django.http import JsonResponse
+from django.shortcuts import render, redirect
 
-# Create your views here.
 from wood.models import *
 
 
@@ -55,6 +57,8 @@ def edit_category(request, category_id):
 
 def delete_category(request, category_id):
     category = Category.objects.get(pk=category_id)
+    if Product.objects.filter(category=category).exists():
+        return error(request, 'Невозможно удалить категорию в которой есть продукты')
     category.image.delete(save=True)
     category.delete()
     return redirect('get_categories')
@@ -217,24 +221,31 @@ def registration(request):
     if request.method == 'GET':
         return render(request, 'create_user.html')
     else:
-        user = User.objects.create_user(username=request.POST['username'],
-                                        email=request.POST['email'],
-                                        password=request.POST['password'],
-                                        first_name=request.POST['first_name'],
-                                        last_name=request.POST['last_name'],
-                                        )
-        user.save()
+        try:
+            user = User.objects.create_user(username=request.POST['username'],
+                                            email=request.POST['email'],
+                                            password=request.POST['password'],
+                                            first_name=request.POST['first_name'],
+                                            last_name=request.POST['last_name'],
+                                            )
+            user.save()
+        except IntegrityError:
+            return error(request, 'Пользователь с таким никнеймом уже существует')
         client = Client(user=user,
                         telephone=request.POST['telephone'],
                         skype=request.POST['skype'],
                         postcode=request.POST['postcode'],
                         address=request.POST['address'])
+
         client.save()
         return redirect('index')
 
 
 def login_user(request):
-    user = authenticate(username=request.POST['username'], password=request.POST['password'])
+    try:
+        user = authenticate(username=request.POST['username'], password=request.POST['password'])
+    except AttributeError:
+        return error(request, 'Неверный пароль')
     if user.is_active:
         login(request, user)
     return redirect('index')
@@ -264,7 +275,10 @@ def view_materials(request):
 
 def delete_material(request, material_id):
     material = Material.objects.get(pk=material_id)
-    material.delete()
+    try:
+        material.delete()
+    except ProtectedError:
+        return error(request, 'Удаляемый материал используется для изделий')
     return redirect('get_materials')
 
 
@@ -289,7 +303,10 @@ def view_coatings(request):
 
 def delete_coatings(request, coating_id):
     coating = Coating.objects.get(pk=coating_id)
-    coating.delete()
+    try:
+        coating.delete()
+    except ProtectedError:
+        return error(request, 'Удаляемое покрытие используется для изделий')
     return redirect('get_coatings')
 
 
@@ -313,7 +330,10 @@ def get_sizes(request):
 
 def delete_size(request, size_id):
     size = Size.objects.get(pk=size_id)
-    size.delete()
+    try:
+        size.delete()
+    except ProtectedError:
+        return error(request, 'Удаляемый размер используется для изделий')
     return redirect('get_sizes')
 
 
@@ -352,7 +372,10 @@ def change_password(request):
             user = authenticate(username=request.user.username, password=request.POST['new_password'])
             if user.is_active:
                 login(request, user)
-
             return HttpResponse("OK")
     else:
         return HttpResponse(status=400)
+
+
+def error(request, error_message):
+    return render(request, 'error_page.html', context={'error_message': error_message})
